refactor: extract ActionImage widget to shared file
Consolidate duplicate action image logic from skill_milestones_dialog.dart
and spend_mastery_dialog.dart into a shared action_image.dart file.
This eliminates ~50 lines of duplicated code.
